pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = { extras = ["server"], version = ">=0.0.30" }
pydantic = "<2"
pirate-speak = { path = "packages/pirate-speak", develop = true }
langchain-openai = "^0.0.2.post1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
